Create README.md
This is a simple to-do list application built using JavaScript and HTML. The application allows users to add tasks to a list, edit the tasks, toggle tasks between completed and incomplete status, and delete all tasks at once.

The application uses the localStorage API to persist the to-do list data across browser sessions. When the page is loaded, the application retrieves the to-do list data from localStorage and displays it on the page. When a user adds a new task, edits an existing task, toggles a task's status, or deletes all tasks, the application updates the to-do list data in localStorage.

The application consists of the following main components:

todo: This is an array that stores the to-do list data. Each item in the array is an object that has two properties: text (the task description) and disabled (a boolean flag that indicates whether the task is completed or not).
todoInput: This is a text input field where users can enter new tasks.
todoList: This is an HTML element (a <ul> tag) where the application displays the list of tasks.
todoCount: This is an HTML element (a <p> tag) that displays the number of tasks in the to-do list.
addButton: This is an HTML button that users can click to add a new task to the to-do list.
deleteButton: This is an HTML button that users can click to delete all tasks from the to-do list.
The application provides the following functionality:

Adding a new task: Users can enter a new task in the todoInput field and click the addButton button to add the task to the to-do list. The application validates the input to ensure that it is not empty.
Editing a task: Users can click an existing task in the todoList to edit it. The task is replaced with an input field where users can enter the updated task description. When the input field loses focus, the application updates the task description in the todo array and saves the updated data to localStorage.
Toggling a task's status: Users can click the checkbox next to a task to toggle its completed status. The application updates the disabled property of the corresponding todo object and saves the updated data to localStorage.
Deleting all tasks: Users can click the deleteButton button to delete all tasks from the to-do list. The application clears the todo array and saves the updated data to localStorage.
Overall, this is a simple and easy-to-use to-do list application that allows users to manage their tasks efficiently.